// src/utils/types.ts
export type Jboss8Status = 'Running' | 'Stopped' | 'Error';

export interface Jboss8Inventory {
  id: number;
  environment: 'Prod' | 'Test' | 'Dev' | 'QA' | 'Lab' | 'Edu';
  server: string;
  instanceName: string;
  version: string;
  status: Jboss8Status;
  port: number;
  jvmVersion: string;
  managementPort?: number;
  heapSize?: string;
  ownerTeam: string;
  lastUpdate: string;
}